+++ crates/wdl-ast/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.23.0 - 2026-05-01
+
 #### Added
 
 * Support for the WDL 1.4 three-form import syntax, including new